+++ src/main/java/com/citytemperature/domain/contract/CityTemperature.java
@@ -2,6 +2,7 @@ package com.citytemperature.domain.contract;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface CityTemperature {
@@ -13,8 +14,8 @@ public interface CityTemperature {
     LocalDate getDateThisTemperatureIsExpected();
 
     @JsonProperty("temperatureInCelsius")
-    Double getTemperatureInCelsius();
+    BigDecimal getTemperatureInCelsius();
 
     @JsonProperty("temperatureInFahrenheit")
-    Double getTemperatureInFahrenheit();
+    BigDecimal getTemperatureInFahrenheit();
 }
